refactor: move Scene into own component
